Hide non-standard APIs from java.util.stream
Review-Link: https://gwt-review.googlesource.com/#/c/17540/
